--- src/Mobile_Pages/Mobile_BlogPage.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Mobile_BlogPage() {
-  return (
-    <div>Mobile_BlogPage</div>
-  )
-}
-
-export default Mobile_BlogPage
