+++ src/Mobile_Pages/Mobile_BusinessBlogPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Mobile_BusinessBlogPage() {
+  return (
+    <div>Mobile_BusinessBlogPage</div>
+  )
+}
+
+export default Mobile_BusinessBlogPage
