+++ src/Mobile_Pages/Mobile_CloudBlogPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Mobile_CloudBlogPage() {
+  return (
+    <div>Mobile_CloudBlogPage</div>
+  )
+}
+
+export default Mobile_CloudBlogPage
